feat: 제네릭
제네릭을 사용해 피연산자가 여러 타입이어도 연산되도록 코드 수정

+++ src/lv3/App.java
@@ -8,18 +8,15 @@ public class App {
         ArithmeticCalculator<Double> calculator = new ArithmeticCalculator<>();
 
         System.out.print("첫 번째 정수를 입력하세요: ");
-        double firstNumber = sc.nextInt();
+        double firstNumber = sc.nextDouble();
 
         System.out.print("두 번째 정수를 입력하세요: ");
-        double secondNumber = sc.nextInt();
+        double secondNumber = sc.nextDouble();
 
         System.out.print("사칙연산 기호를 입력하세요: ");
         char operator = sc.next().charAt(0);
-
-        // enum 클래스의 getType() 메서드를 통해 입력 값이 열거 상수의 값과 일치하는지 판단하고 operatorType 변수에 저장
         OperatorType operatorType = OperatorType.getType(operator);
-
-        // 테스트
+        
         double result = calculator.calculate(firstNumber, secondNumber, operatorType);
         System.out.println("result = " + result);
     }
